change int conversion in motor_controller to RPM velocity and change some lines in movement_manager to perform some tests

+++ src/motor_controller.py
@@ -30,8 +30,8 @@ def compute_rpms(data):
     #v_left_wheel = vx - wz*d/2
     v_left_wheel = vx/2 - wz*d/4
 
-    rpm_right_wheel = int(v_right_wheel*60/(r*2*np.pi))
-    rpm_left_wheel = int(v_left_wheel*60/(r*2*np.pi))
+    rpm_right_wheel = v_right_wheel*60/(r*2*np.pi)
+    rpm_left_wheel = v_left_wheel*60/(r*2*np.pi)
 
     rospy.loginfo(rpm_right_wheel)
     rospy.loginfo(rpm_left_wheel)
